reverse-search/requirements.txt
python-dotenv
google-cloud-storage
google-cloud-vision
openai
numpy
scipy
scikit-learn
Pillow
sanic
requests